#118: Moved search history query parameter methods
from model to separate file search_history.py
