+++ signbank/dictionary/models.py
@@ -3080,120 +3080,6 @@ class SearchHistory(models.Model):
         query_languages = list(set(language_parameters))
         return query_languages
 
-    def display_parameters(self):
-        # this method displays the parameters as a string, using ** to separate each
-        # for the multi-select fields the choices are shown together
-        parameter_dict = dict()
-        # for the language search fields, this dictionary is used to group them together
-        # so searches on annotation or lemma or keyword fields appear next to each other
-        # this groups them by language
-        parameters_multilingual_dict = dict()
-        for translation in self.parameters.all():
-            if translation.is_fieldchoice():
-                field_choice = translation.queryparameterfieldchoice
-                field_name = field_choice.display_verbose_fieldname()
-                if field_name not in parameter_dict.keys():
-                    parameter_dict[field_name] = []
-                parameter_dict[field_name].append(field_choice.fieldValue.name)
-            elif translation.is_handshape():
-                handshape = translation.queryparameterhandshape
-                field_name = handshape.display_verbose_fieldname()
-                if field_name not in parameter_dict.keys():
-                    parameter_dict[field_name] = []
-                parameter_dict[field_name].append(handshape.fieldValue.name)
-            elif translation.is_semanticfield():
-                semanticfield = translation.queryparametersemanticfield
-                field_name = semanticfield.display_verbose_fieldname()
-                if field_name not in parameter_dict.keys():
-                    parameter_dict[field_name] = []
-                parameter_dict[field_name].append(semanticfield.fieldValue.name)
-            elif translation.is_derivationhistory():
-                derivationhistory = translation.queryparameterderivationhistory
-                field_name = derivationhistory.display_verbose_fieldname()
-                if field_name not in parameter_dict.keys():
-                    parameter_dict[field_name] = []
-                parameter_dict[field_name].append(derivationhistory.fieldValue.name)
-            elif translation.is_boolean():
-                nullbooleanfield = translation.queryparameterboolean
-                field_name = nullbooleanfield.display_verbose_fieldname()
-                if field_name not in parameter_dict.keys():
-                    parameter_dict[field_name] = []
-                parameter_dict[field_name].append(str(nullbooleanfield.fieldValue))
-            elif translation.is_multilingual():
-                multilingual = translation.queryparametermultilingual
-                field_name = multilingual.fieldName
-                field_name_verbose = multilingual.display_verbose_fieldname()
-                if field_name not in parameters_multilingual_dict.keys():
-                    parameters_multilingual_dict[field_name] = dict()
-                if field_name_verbose not in parameters_multilingual_dict[field_name].keys():
-                    parameters_multilingual_dict[field_name][field_name_verbose] = []
-                parameters_multilingual_dict[field_name][field_name_verbose].append(multilingual.fieldValue)
-            else:
-                pass
-        # now add the grouped language search fields
-        for search_field in parameters_multilingual_dict.keys():
-            for verbose_field in parameters_multilingual_dict[search_field].keys():
-                parameter_dict[verbose_field] = parameters_multilingual_dict[search_field][verbose_field]
-
-        parameters_string = " ** ".join(key + ": " + ", ".join(values) for (key, values) in parameter_dict.items())
-        return parameters_string
-
-    def query_parameters(self):
-        # this method converts the search parameters to match the format of the query parameters in the session
-        search_history_parameters = dict()
-        sh_parameters = self.parameters.all()
-        for shp in sh_parameters:
-            if shp.is_fieldchoice():
-                field_choice = shp.queryparameterfieldchoice
-                field_name = field_choice.fieldName + '[]'
-                if field_name not in search_history_parameters.keys():
-                    search_history_parameters[field_name] = []
-                field_machine_value = field_choice.fieldValue.machine_value if field_choice.fieldValue else 0
-                search_history_parameters[field_name].append(str(field_machine_value))
-            elif shp.is_handshape():
-                handshape = shp.queryparameterhandshape
-                field_name = handshape.fieldName + '[]'
-                if field_name not in search_history_parameters.keys():
-                    search_history_parameters[field_name] = []
-                field_machine_value = handshape.fieldValue.machine_value if handshape.fieldValue else 0
-                search_history_parameters[field_name].append(str(field_machine_value))
-            elif shp.is_semanticfield():
-                semanticfield = shp.queryparametersemanticfield
-                field_name = semanticfield.fieldName + '[]'
-                if field_name not in search_history_parameters.keys():
-                    search_history_parameters[field_name] = []
-                field_machine_value = semanticfield.fieldValue.machine_value if semanticfield.fieldValue else 0
-                search_history_parameters[field_name].append(str(field_machine_value))
-            elif shp.is_derivationhistory():
-                derivationhistory = shp.queryparameterderivationhistory
-                field_name = derivationhistory.fieldName + '[]'
-                if field_name not in search_history_parameters.keys():
-                    search_history_parameters[field_name] = []
-                field_machine_value = derivationhistory.fieldValue.machine_value if derivationhistory.fieldValue else 0
-                search_history_parameters[field_name].append(str(field_machine_value))
-            elif shp.is_boolean():
-                nullbooleanfield = shp.queryparameterboolean
-                field_name = nullbooleanfield.fieldName
-                if field_name not in search_history_parameters.keys():
-                    search_history_parameters[field_name] = []
-                NEUTRALBOOLEANCHOICES = {'None': '1', 'True': '2', 'False': '3'}
-                YESNOCHOICES = {'None': 'unspecified', 'True': 'yes', 'False': 'no'}
-                RELATIONCHOICES = {'None': '0', 'True': '1', 'False': '2'}
-                if field_name == 'defspublished':
-                    field_value = YESNOCHOICES[str(nullbooleanfield.fieldValue)]
-                elif field_name == 'hasRelationToForeignSign':
-                    field_value = RELATIONCHOICES[str(nullbooleanfield.fieldValue)]
-                else:
-                    field_value = NEUTRALBOOLEANCHOICES[str(nullbooleanfield.fieldValue)]
-                search_history_parameters[field_name] = field_value
-            elif shp.is_multilingual():
-                multilingual = shp.queryparametermultilingual
-                field_name = multilingual.fieldName + '_' + multilingual.fieldLanguage.language_code_2char
-                if field_name not in search_history_parameters.keys():
-                    search_history_parameters[field_name] = []
-                search_history_parameters[field_name] = multilingual.fieldValue
-        return search_history_parameters
-
 
 CATEGORY_MODELS_MAPPING = {
     'SemField': SemanticField,
