+++ signbank/query_parameters.py
@@ -30,136 +30,6 @@ from django.shortcuts import render, get_object_or_404, redirect
 from guardian.shortcuts import get_objects_for_user
 
 
-def available_query_parameters_in_search_history():
-    available_parameters = []
-    available_subclasses = QueryParameter.__subclasses__()
-    for classmodel in available_subclasses:
-        if hasattr(classmodel, 'QUERY_FIELDS'):
-            available_parameters = available_parameters + classmodel.QUERY_FIELDS
-    parameters = [t1 for (t1, t2) in available_parameters]
-    return parameters
-
-
-def languages_in_query(queryid):
-    query = get_object_or_404(SearchHistory, id=queryid)
-    multiparameters_of_query = QueryParameterMultilingual.objects.filter(search_history=query)
-    languages = multiparameters_of_query.values('fieldLanguage')
-    languageids = [lang['fieldLanguage'] for lang in languages]
-    language_objects = Language.objects.filter(id__in=languageids).distinct()
-    return language_objects
-
-
-def fieldnames_from_query_parameters(query_parameters):
-    glosssearch = "glosssearch_"
-    lemmasearch = "lemma_"
-    keywordsearch = "keyword_"
-
-    fieldnames = []
-    for key in query_parameters:
-        if key == 'search_type':
-            continue
-        elif key[-2:] == '[]':
-            fieldnames.append(key[:-2])
-        elif key.startswith(glosssearch) or key.startswith(lemmasearch) or key.startswith(keywordsearch):
-            fieldnames.append(key[:-3])
-        else:
-            fieldnames.append(key)
-    return fieldnames
-
-
-def save_query_parameters(request, query_name, query_parameters):
-
-    glosssearch = "glosssearch_"
-    lemmasearch = "lemma_"
-    keywordsearch = "keyword_"
-
-    search_history = SearchHistory(user=request.user, queryName=query_name)
-    search_history.save()
-    for key in query_parameters:
-        if key == 'search_type':
-            continue
-        elif key[-2:] == '[]':
-            # multiple choice fields have a list of values
-            if key[:-2] in ['domhndsh', 'subhndsh', 'final_domhndsh', 'final_subhndsh']:
-                choices_for_category = Handshape.objects.filter(machine_value__in=query_parameters[key])
-                for query_value in choices_for_category:
-                    qp = QueryParameterHandshape(fieldName=key[:-2], fieldValue=query_value, search_history=search_history)
-                    qp.save()
-                    search_history.parameters.add(qp)
-            elif key[:-2] in ['semField']:
-                choices_for_category = SemanticField.objects.filter(machine_value__in=query_parameters[key])
-                for query_value in choices_for_category:
-                    qp = QueryParameterSemanticField(fieldName=key[:-2], fieldValue=query_value, search_history=search_history)
-                    qp.save()
-                    search_history.parameters.add(qp)
-            elif key[:-2] in ['derivHist']:
-                choices_for_category = DerivationHistory.objects.filter(machine_value__in=query_parameters[key])
-                for query_value in choices_for_category:
-                    qp = QueryParameterDerivationHistory(fieldName=key[:-2], fieldValue=query_value, search_history=search_history)
-                    qp.save()
-                    search_history.parameters.add(qp)
-            else:
-                field = key[:-2]
-                field_category = Gloss._meta.get_field(field).field_choice_category
-                choices_for_category = FieldChoice.objects.filter(field__iexact=field_category, machine_value__in=query_parameters[key])
-                for query_value in choices_for_category:
-                    qp = QueryParameterFieldChoice(fieldName=key[:-2], fieldValue=query_value, search_history=search_history)
-                    qp.save()
-                    search_history.parameters.add(qp)
-        elif key in ['weakdrop', 'weakprop', 'domhndsh_letter', 'domhndsh_number',
-                     'subhndsh_letter', 'subhndsh_number', 'repeat', 'altern', 'inWeb', 'isNew',
-                     'excludeFromEcv', 'hasvideo', 'hasothermedia', 'defspublished', 'hasRelationToForeignSign']:
-            NEUTRALBOOLEANCHOICES = {'0': None, '1': None, '2': True, '3': False}
-            UNKNOWNBOOLEANCHOICES = {'0': False, '2': True, '3': False}
-            if key in ['weakdrop', 'weakprop']:
-                query_value = NEUTRALBOOLEANCHOICES[query_parameters[key]]
-            elif key in ['domhndsh_letter', 'domhndsh_number', 'subhndsh_letter', 'subhndsh_number']:
-                query_value = UNKNOWNBOOLEANCHOICES[query_parameters[key]]
-            elif key in ['repeat', 'altern']:
-                query_value = UNKNOWNBOOLEANCHOICES[query_parameters[key]]
-            elif key in ['defspublished']:
-                query_value = query_parameters[key] == 'yes'
-            elif key in ['hasRelationToForeignSign']:
-                query_value = query_parameters[key] == '1'
-            else:
-                # inWeb, isNew , excludeFromEcv, hasvideo, hasothermedia
-                query_value = query_parameters[key] == '2'
-            qp = QueryParameterBoolean(fieldName=key, fieldValue=query_value,
-                                       search_history=search_history, multiselect=False)
-            qp.save()
-            search_history.parameters.add(qp)
-        elif key.startswith(glosssearch):
-            search_field = glosssearch[:-1]
-            search_value = query_parameters[key]
-            language_code_2char = key[len(glosssearch):]
-            language = Language.objects.get(language_code_2char=language_code_2char)
-            qp = QueryParameterMultilingual(fieldName=search_field, fieldLanguage=language,
-                                            fieldValue=search_value, search_history=search_history, multiselect=False)
-            qp.save()
-            search_history.parameters.add(qp)
-        elif key.startswith(lemmasearch):
-            search_field = lemmasearch[:-1]
-            search_value = query_parameters[key]
-            language_code_2char = key[len(lemmasearch):]
-            language = Language.objects.get(language_code_2char=language_code_2char)
-            qp = QueryParameterMultilingual(fieldName=search_field, fieldLanguage=language,
-                                            fieldValue=search_value, search_history=search_history, multiselect=False)
-            qp.save()
-            search_history.parameters.add(qp)
-        elif key.startswith(keywordsearch):
-            search_field = keywordsearch[:-1]
-            search_value = query_parameters[key]
-            language_code_2char = key[len(keywordsearch):]
-            language = Language.objects.get(language_code_2char=language_code_2char)
-            qp = QueryParameterMultilingual(fieldName=search_field, fieldLanguage=language,
-                                            fieldValue=search_value, search_history=search_history, multiselect=False)
-            qp.save()
-            search_history.parameters.add(qp)
-        else:
-            continue
-    search_history.save()
-
-
 def list_to_query(query_list):
     # this ANDs together the individual Q elements in the query_list
     if not query_list:
